fix(build_system): Move job functions to sdcm.utils.ci_tools
