+++ sdcm/argus_test_run.py
@@ -15,7 +15,7 @@ from sdcm.sct_config import SCTConfiguration
 from sdcm.utils.common import get_test_name, get_sct_runner_ip, get_my_ip
 from sdcm.utils.get_username import get_username
 from sdcm.utils.git import get_git_current_branch, get_git_commit_id
-from utils.build_system import get_job_url, get_job_name
+from sdcm.utils.ci_tools import get_job_url, get_job_name
 
 LOGGER = logging.getLogger(__name__)
