+++ sdcm/db_stats.py
@@ -35,7 +35,7 @@ from sdcm.utils.common import normalize_ipv6_url
 from sdcm.utils.git import get_git_commit_id
 from sdcm.utils.decorators import retrying
 from sdcm.sct_events.system import ElasticsearchEvent
-from utils.build_system import get_job_name, get_job_url
+from sdcm.utils.ci_tools import get_job_name, get_job_url
 
 LOGGER = logging.getLogger(__name__)
